backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
colorama==0.4.6
dj-database-url==1.2.0
dj-static==0.0.6
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==16.6.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
static3==0.7.0
tzdata==2022.7
